backend/requirements.txt
fastapi
uvicorn[standard]
google-genai
pytest
pytest-asyncio
pydantic-settings
python-dotenv
httpx
